requirements.txt
SQLAlchemy==1.2.6
sortedcontainers==1.5.9
avro-python3==1.8.2
dsfaker==0.2.2
Flask==0.12
Flask-RESTful==0.3.6